Rename embed -> ruleset
